backend/requirements.txt
alembic==1.8.1
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.18.0
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
jsonschema==3.2.0
Mako==1.2.2
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.12.0
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
texttable==1.6.7
typing_extensions==4.3.0
urllib3==1.26.13
uvicorn==0.18.3
websocket-client==0.59.0
websockets==10.3
